requirements.txt
Twisted==24.3.0
klein==23.12.0
txpostgres==1.7.0
psycopg2-binary==2.9.9
PyYAML==6.0.2
pydantic==2.8.2
Jinja2==3.1.4
python-dotenv==1.0.1
